refactor modules and import locs
