+++ TODO
@@ -2,8 +2,6 @@
 
 ## Top
 
--   [x] nits -> change base_blocks to just blocks in st_flow, compute and other places
--   [ ] nits -> change flow.flow.types to flow.types
 -   [ ] test optional dependencies
 -   [ ] ruff checks & release 1.0.0
 
@@ -66,7 +64,8 @@
 -   [x] update changelog
 -   [x] update dev notes
 -   [x] update docs for the images and other pages
--   [ ] nits -> change base_blocks to just blocks in st_flow, compute and other places
--   [ ] nits -> change flow.flow.types to flow.types
+-   [x] nits -> change base_blocks to just blocks in st_flow, compute and other places
+-   [x] nits -> change flow.flow.types to flow.types
+-   [x] move st_flow from base to st along with response types
 -   [ ] test optional dependencies
 -   [ ] ruff checks & release 1.0.0
