+++ src/barfi/flow/__init__.py
@@ -1,4 +1,4 @@
 from barfi.flow.block import Block as Block
 from barfi.flow.schema import SchemaManager as SchemaManager
 from barfi.flow.compute import ComputeEngine as ComputeEngine
-from barfi.flow.base import st_flow as st_flow
+from barfi.flow.streamlit import st_flow as st_flow
